+++ processor/src/test/resources/expected/com/example/observer_ref/Arez_CustomNameRefOnObserveModel3.java
@@ -1,0 +1,119 @@
+package com.example.observer_ref;
+
+import arez.Arez;
+import arez.ArezContext;
+import arez.Component;
+import arez.Disposable;
+import arez.Flags;
+import arez.Observer;
+import arez.component.DisposeNotifier;
+import arez.component.DisposeTrackable;
+import arez.component.Identifiable;
+import arez.component.internal.ComponentKernel;
+import javax.annotation.Generated;
+import javax.annotation.Nonnull;
+import org.realityforge.braincheck.Guards;
+
+@Generated("arez.processor.ArezProcessor")
+public final class Arez_CustomNameRefOnObserveModel3 extends CustomNameRefOnObserveModel3 implements Disposable, Identifiable<Integer>, DisposeTrackable {
+  private static volatile int $$arezi$$_nextId;
+
+  private final ComponentKernel $$arezi$$_kernel;
+
+  @Nonnull
+  private final Observer $$arez$$_render;
+
+  public Arez_CustomNameRefOnObserveModel3() {
+    super();
+    final ArezContext $$arezv$$_context = Arez.context();
+    final int $$arezv$$_id = ( Arez.areNamesEnabled() || Arez.areRegistriesEnabled() || Arez.areNativeComponentsEnabled() ) ? ++$$arezi$$_nextId : 0;
+    final String $$arezv$$_name = Arez.areNamesEnabled() ? "CustomNameRefOnObserveModel3." + $$arezv$$_id : null;
+    final Component $$arezv$$_component = Arez.areNativeComponentsEnabled() ? $$arezv$$_context.component( "CustomNameRefOnObserveModel3", $$arezv$$_id, $$arezv$$_name, () -> $$arezi$$_nativeComponentPreDispose() ) : null;
+    this.$$arezi$$_kernel = new ComponentKernel( Arez.areZonesEnabled() ? $$arezv$$_context : null, $$arezv$$_name, $$arezv$$_id, $$arezv$$_component, null, Arez.areNativeComponentsEnabled() ? null : this::$$arezi$$_dispose, null, true, false, false );
+    this.$$arez$$_render = $$arezv$$_context.tracker( Arez.areNativeComponentsEnabled() ? $$arezv$$_component : null, Arez.areNamesEnabled() ? $$arezv$$_name + ".render" : null, () -> super.onRenderDepsChan$$$$$ge(), Flags.RUN_LATER | Flags.NESTED_ACTIONS_DISALLOWED | Flags.AREZ_DEPENDENCIES | Flags.ENVIRONMENT_NOT_REQUIRED );
+    this.$$arezi$$_kernel.componentConstructed();
+    this.$$arezi$$_kernel.componentComplete();
+  }
+
+  final int $$arezi$$_id() {
+    return this.$$arezi$$_kernel.getId();
+  }
+
+  @Override
+  @Nonnull
+  public final Integer getArezId() {
+    return $$arezi$$_id();
+  }
+
+  private void $$arezi$$_nativeComponentPreDispose() {
+    this.$$arezi$$_kernel.getDisposeNotifier().dispose();
+  }
+
+  @Override
+  @Nonnull
+  public DisposeNotifier getNotifier() {
+    return this.$$arezi$$_kernel.getDisposeNotifier();
+  }
+
+  @Override
+  public boolean isDisposed() {
+    return this.$$arezi$$_kernel.isDisposed();
+  }
+
+  @Override
+  public void dispose() {
+    this.$$arezi$$_kernel.dispose();
+  }
+
+  private void $$arezi$$_dispose() {
+    this.$$arez$$_render.dispose();
+  }
+
+  @Override
+  public void $$$rende$$$r() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named '$$$rende$$$r' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? '?' : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    this.$$arezi$$_kernel.getContext().safeObserve( this.$$arez$$_render, () -> super.$$$rende$$$r(), null );
+  }
+
+  @Override
+  Observer observe$$$r() {
+    if ( Arez.shouldCheckApiInvariants() ) {
+      Guards.apiInvariant( () -> null != this.$$arezi$$_kernel && this.$$arezi$$_kernel.isActive(), () -> "Method named 'observe$$$r' invoked on " + this.$$arezi$$_kernel.describeState() + " component named '" + ( null == this.$$arezi$$_kernel ? '?' : this.$$arezi$$_kernel.getName() ) + "'" );
+    }
+    return $$arez$$_render;
+  }
+
+  @Override
+  public final int hashCode() {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      return Integer.hashCode( $$arezi$$_id() );
+    } else {
+      return super.hashCode();
+    }
+  }
+
+  @Override
+  public final boolean equals(final Object o) {
+    if ( Arez.areNativeComponentsEnabled() ) {
+      if ( o instanceof Arez_CustomNameRefOnObserveModel3 ) {
+        final Arez_CustomNameRefOnObserveModel3 that = (Arez_CustomNameRefOnObserveModel3) o;
+        return $$arezi$$_id() == that.$$arezi$$_id();
+      } else {
+        return false;
+      }
+    } else {
+      return super.equals( o );
+    }
+  }
+
+  @Override
+  public final String toString() {
+    if ( Arez.areNamesEnabled() ) {
+      return "ArezComponent[" + this.$$arezi$$_kernel.getName() + "]";
+    } else {
+      return super.toString();
+    }
+  }
+}