+++ processor/src/test/resources/input/com/example/action/NonStandardNameActionModel.java
@@ -1,0 +1,13 @@
+package com.example.action;
+
+import arez.annotations.Action;
+import arez.annotations.ArezComponent;
+
+@ArezComponent
+public abstract class NonStandardNameActionModel
+{
+  @Action
+  public void $$my$Funky$Action$$_( final long $$time, float someOtherParameter$$ )
+  {
+  }
+}