+++ processor/src/test/resources/input/com/example/id/NonStandardNameModel.java
@@ -1,0 +1,14 @@
+package com.example.id;
+
+import arez.annotations.ArezComponent;
+import arez.annotations.ComponentId;
+
+@ArezComponent( allowEmpty = true )
+abstract class NonStandardNameModel
+{
+  @ComponentId
+  final int getId()
+  {
+    return 0;
+  }
+}