+++ CHANGELOG.md
@@ -6,6 +6,8 @@
   be executed the Arez scheduler. If the scheduler is currently executing tasks, the task will be added
   to the queue and executed in turn otherwise the scheduler will be activated. This functionality is primarily
   used to enable better integration between the Arez runtime and external libraries.
+* **\[processor\]** Fix bug where methods annotated with `@Action` or `@Observe` could not contain `$`
+  characters. Add tests to ensure other Arez annotated methods and/or fields support `$` in the name.  
 
 ### [v0.116](https://github.com/arez/arez/tree/v0.116) (2018-11-21)
 [Full Changelog](https://github.com/arez/arez/compare/v0.115...v0.116)